prevent the user to remove its account, add search by transaction id and user in teh by the same search input

+++ src/controllers/UserController.ts
@@ -87,6 +87,11 @@ export default class UserController {
   ) => {
     try {
       const { id } = req.params;
+      const { user } = res.locals;
+
+      if (user.id == id) {
+        throw new GenericError(`User can't remove it's own account`, 400);
+      }
 
       await UserQuery.deleteOneUser(Number(id));
